Device/Declaration: size() returns unsigned, not size_t

+++ src/Device/Declaration.hpp
@@ -60,7 +60,7 @@ struct Declaration {
     return waypoints[i].Location;
   }
 
-  size_t size() const {
+  unsigned size() const {
     return waypoints.size();
   }
 };
